v3 - Splitting Codebase - ConfigurableAreas, CollapsingColumns and  TableAttributes (#1514)
* Splitting ConfigurableAreas into Traits

* Splitting TableAttributes into Traits

* Splitting CollapsingColumns into Traits

* Add HasAllTraits Trait


---------
