+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 All notable changes to `laravel-livewire-tables` will be documented in this file
 
 ## UNRELEASED
-- Split ConfigurableArea into own Trait/Config/Helper Files
+- Split ConfigurableAreas, CollapsingColumns and  TableAttributes into own Traits/Config/Helper Files for Maintainability by @lrljoe in [#1514](https://github.com/rappasoft/laravel-livewire-tables/pull/1514)
+- Add "HasAllTraits" for Maintainability by @lrljoe in [#1514](https://github.com/rappasoft/laravel-livewire-tables/pull/1514)
 
 ## [v3.1.0] - 2023-10-31
 - Restore wire:confirm for Bulk Actions
