+++ src/Traits/Configuration/CollapsingColumnConfiguration.php
@@ -1,0 +1,27 @@
+<?php
+
+namespace Rappasoft\LaravelLivewireTables\Traits\Configuration;
+
+trait CollapsingColumnConfiguration
+{
+    public function setCollapsingColumnsStatus(bool $status): self
+    {
+        $this->collapsingColumnsStatus = $status;
+
+        return $this;
+    }
+
+    public function setCollapsingColumnsEnabled(): self
+    {
+        $this->setCollapsingColumnsStatus(true);
+
+        return $this;
+    }
+
+    public function setCollapsingColumnsDisabled(): self
+    {
+        $this->setCollapsingColumnsStatus(false);
+
+        return $this;
+    }
+}
