+++ src/Traits/WithTableAttributes.php
@@ -1,0 +1,34 @@
+<?php
+
+namespace Rappasoft\LaravelLivewireTables\Traits;
+
+use Rappasoft\LaravelLivewireTables\Traits\Configuration\TableAttributeConfiguration;
+use Rappasoft\LaravelLivewireTables\Traits\Helpers\TableAttributeHelpers;
+
+trait WithTableAttributes
+{
+    use TableAttributeConfiguration,
+        TableAttributeHelpers;
+
+    protected array $componentWrapperAttributes = [];
+
+    protected array $tableWrapperAttributes = [];
+
+    protected array $tableAttributes = [];
+
+    protected array $theadAttributes = [];
+
+    protected array $tbodyAttributes = [];
+
+    protected $thAttributesCallback;
+
+    protected $thSortButtonAttributesCallback;
+
+    protected $trAttributesCallback;
+
+    protected $tdAttributesCallback;
+
+    protected $trUrlCallback;
+
+    protected $trUrlTargetCallback;
+}
